Compute fn registry
