+++ encodings/fsst/src/array.rs
@@ -25,6 +25,7 @@ pub struct FSSTArray {
     stats_set: ArrayStats,
 }
 
+#[derive(Debug)]
 pub struct FSSTEncoding;
 impl Encoding for FSSTEncoding {
     type Array = FSSTArray;
