+++ encodings/fsst/src/compute/mod.rs
@@ -1,16 +1,13 @@
 mod compare;
+mod filter;
 
 use vortex_array::arrays::{VarBinArray, varbin_scalar};
 use vortex_array::builders::ArrayBuilder;
-use vortex_array::compute::{
-    FilterKernel, FilterKernelAdapter, ScalarAtFn, SliceFn, TakeFn, fill_null, filter, scalar_at,
-    slice, take,
-};
+use vortex_array::compute::{ScalarAtFn, SliceFn, TakeFn, fill_null, scalar_at, slice, take};
 use vortex_array::vtable::ComputeVTable;
-use vortex_array::{Array, ArrayExt, ArrayRef, register_kernel};
+use vortex_array::{Array, ArrayExt, ArrayRef};
 use vortex_buffer::ByteBuffer;
 use vortex_error::{VortexResult, vortex_err};
-use vortex_mask::Mask;
 use vortex_scalar::{Scalar, ScalarValue};
 
 use crate::{FSSTArray, FSSTEncoding};
@@ -88,19 +85,3 @@ impl ScalarAtFn<&FSSTArray> for FSSTEncoding {
         Ok(varbin_scalar(decoded_buffer, array.dtype()))
     }
 }
-
-impl FilterKernel for FSSTEncoding {
-    // Filtering an FSSTArray filters the codes array, leaving the symbols array untouched
-    fn filter(&self, array: &FSSTArray, mask: &Mask) -> VortexResult<ArrayRef> {
-        Ok(FSSTArray::try_new(
-            array.dtype().clone(),
-            array.symbols().clone(),
-            array.symbol_lengths().clone(),
-            filter(array.codes(), mask)?.as_::<VarBinArray>().clone(),
-            filter(array.uncompressed_lengths(), mask)?,
-        )?
-        .into_array())
-    }
-}
-
-register_kernel!(FilterKernelAdapter(FSSTEncoding).lift());
