+++ vortex-array/src/arrays/struct_/mod.rs
@@ -27,6 +27,7 @@ pub struct StructArray {
     stats_set: ArrayStats,
 }
 
+#[derive(Debug)]
 pub struct StructEncoding;
 impl Encoding for StructEncoding {
     type Array = StructArray;
