+++ vortex-array/src/encoding.rs
@@ -10,7 +10,7 @@ use crate::{Array, DeserializeMetadata, SerializeMetadata};
 pub type EncodingId = ArcRef<str>;
 
 /// Marker trait for array encodings with their associated Array type.
-pub trait Encoding: 'static + Send + Sync + EncodingVTable {
+pub trait Encoding: 'static + Send + Sync + EncodingVTable + Debug {
     type Array: Array;
     type Metadata: SerializeMetadata + DeserializeMetadata + Debug;
